1.优化bpmn解析 2.数据存储优化

+++ model/historicIdentityLink.go
@@ -4,7 +4,7 @@ type HistoricIdentityLink struct {
 	Id                int64
 	Type              string `gorm:"column:type"`
 	TaskId            int64  `gorm:"column:task_id"`
-	ProcessInstanceId int    `gorm:"column:proc_inst_id"`
+	ProcessInstanceId int64  `gorm:"column:proc_inst_id"`
 	GroupId           string `gorm:"column:group_id"`
 	UserId            string `gorm:"column:user_id"`
 }
